feat: calling exchange-service

+++ src/main/java/dev/felipeazsantos/book_service/model/Book.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 
 import java.io.Serial;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.Objects;
 
@@ -27,7 +28,7 @@ public class Book implements Serializable {
     private Date launchDate;
 
     @Column(nullable = false)
-    private Double price;
+    private BigDecimal price;
 
     @Transient
     private String currency;
@@ -35,7 +36,7 @@ public class Book implements Serializable {
     @Transient
     private String environment;
 
-    public Book(Long id, String author, String title, Date launchDate, Double price, String currency, String environment) {
+    public Book(Long id, String author, String title, Date launchDate, BigDecimal price, String currency, String environment) {
         this.id = id;
         this.author = author;
         this.title = title;
@@ -81,11 +82,11 @@ public class Book implements Serializable {
         this.launchDate = launchDate;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
